+++ lib/phoenix/view.ex
@@ -1,37 +1,100 @@
 defmodule Phoenix.View do
   @moduledoc """
-  Serves as the base view for an entire Phoenix application view layer
+  Defines the view layer of a Phoenix application.
 
-  Users define `App.Views` and `use Phoenix.View`. The main view:
+  This module is used to define the application main view, which
+  serves as the base for all other views and templates in the
+  application.
 
-    * Serves as a base presentation layer for all views and templates
-    * Wires up the Template and template path all for all other views
-    * Expects the base view to define a `__using__` macro for other view modules
+  The view layer also contains conveniences for rendering templates,
+  including support for layours and encoders per format.
 
   ## Examples
 
-      defmodule App.Views do
-        defmacro __using__(_options) do
-          quote do
-            use Phoenix.View, templates_root: unquote(Path.join([__DIR__, "templates"]))
-            import unquote(__MODULE__)
+  First, define the main view module:
 
-            # This block is expanded within all views for aliases, imports, etc
-            def title, do: "Welcome to Phoenix!"
-          end
+      defmodule YourApp.View do
+        use Phoenix.View, root: "web/templates"
+
+        # Everything in this block is available runs in this
+        # module and in other views that use YourApp.View
+        using do
+          # Import common functionality
+          import YourApp.I18n
+          import YourApp.Router.Helpers
+
+          # Use Phoenix.HTML to import all HTML functions (forms, tags, etc)
+          use Phoenix.HTML
         end
 
         # Functions defined here are available to all other views/templates
       end
 
-      defmodule App.PageView
-        use App.Views
+  Once the main view module is defined, we can use it to define each view:
 
-        def display(something) do
-          String.upcase(something)
+      defmodule YourApp.UserView do
+        use YourApp.View
+      end
+
+  Because we have defined the template root to be "web/template", `Phoenix.View`
+  will automatically load all templates at "web/template/user" and include them
+  in the `YourApp.UserView`. For example, imagine we have the template:
+
+      # web/templates/user/index.html.eex
+      Hello <%= @name %>
+
+  The `.eex` extension is called a template engine which tells Phoenix how
+  to compile the code in the file into actual Elixir source code. After compiled,
+  the template can be rendered as:
+
+      Phoenix.View.render(YourApp.UserView, "index.html", name: "John Doe")
+      #=> {:safe, "Hello John Doe"}
+
+  We will discuss rendering with detail next.
+
+  ## Rendering
+
+  The main responsability of a view is to render a template.
+
+  A template has a name, which is also contains a format. For example,
+  in the previous section we have rendered the "index.html" template:
+
+      Phoenix.View.render(YourApp.UserView, "index.html", name: "John Doe")
+      #=> {:safe, "Hello John Doe"}
+
+  When a view renders a template, the result returned is an inner
+  representation specific to the template format. In the example above,
+  we got: `{:safe, "Hello John Doe"}`. The safe tuple annotates that our
+  template is safe and that we don't need to escape its contents because
+  all data was already encoded so far. Let's try to inject custom code:
+
+      Phoenix.View.render(YourApp.UserView, "index.html", name: "John<br />Doe")
+      #=> {:safe, "Hello John&lt;br /&gt;Doe"}
+
+  This inner representation allows us to render and compose templates easily.
+  For example, if you want to render JSON data, we could do so by adding a
+  "show.json" entry to `render/2` in our view:
+
+      defmodule YourApp.UserView do
+        use YourApp.View
+
+        def render("user.json", %{user: user}) do
+          %{name: user.name, address: user.address}
         end
       end
 
+  Notice that, in order to render JSON data, we don't need to explicitly
+  return a JSON string! Instead we just return data that is encodable to
+  JSON.
+
+  Both JSON and HTML formats will be encoded only when passing the data
+  to the controller via the `render_to_iodata/3` function. The
+  `render_to_iodata/3` uses the notion of format encoders to convert a
+  particular format to its string/iodata representation.
+
+  Phoenix ships with some template engines and format encoders, which
+  can be further configured in the Phoenix application. You can read
+  more about format encoders in `Phoenix.Template` documentation.
   """
 
   @doc false
@@ -130,30 +193,41 @@ defmodule Phoenix.View do
   end
 
   @doc """
-  Renders template to String
+  Renders a template.
 
-    * module - The View module, ie, MyView
-    * template - The String template, ie, "index.html"
-    * assigns - The Dictionary of assigns, ie, [title: "Hello!"]
+  It expects the view module, the template as a string, and a
+  set of assigns.
+
+  Notice this function returns the inner representation of a
+  template. If you want the encoded template as a result, use
+  `render_to_iodata/3` instead.
 
   ## Examples
 
-      View.render(MyView, "index.html", title: "Hello!")
-      #=> "<h1>Hello!</h1>"
+      Phoenix.View.render(YourApp.UserView, "index.html", name: "John Doe")
+      #=> {:safe, "Hello John Doe"}
+
+  ## Assigns
+
+  Assigns are meant to be user data that will be available in templates.
+  However, there are keys under assigns that are specially handled by
+  Phoenix, they are:
+
+    * :within - tells Phoenix to wrap the rendered result in the
+      given layout. See next section.
 
   ## Layouts
 
-  Template can be rendered within other templates using the `within` option.
-  `within` accepts a Tuple, of the form `{LayoutModule, "template.extension"}`
+  Template can be rendered within other templates using the `:within`
+  option. :within` accepts a tuple of the form `{LayoutModule, "template.extension"}`.
 
-  When the sub template is rendered, the layout template will have an `@inner`
-  assign containing the rendered contents of the sub-template. For html
-  templates, `@inner` will be passed through `Phoenix.HTML.safe/1` automatically.
+  When a template is rendered, the layout template will have an `@inner`
+  assign containing the rendered contents of the sub-template. For HTML
+  templates, `@inner` will be always marked as safe.
 
-  ### Examples
-
-      View.render(MyView, "index.html", within: {LayoutView, "app.html"})
-      #=> "<html><h1>Hello!</h1></html>"
+      Phoenix.View.render(YourApp.UserView, "index.html",
+                          within: {YourApp.LayoutView, "application.html"})
+      #=> {;safe, "<html><h1>Hello!</h1></html>"}
 
   """
   def render(module, template, assigns) do